flask-jwt-auth/requirements.txt
alembic==0.8.9
bcrypt==3.1.2
certifi==2020.12.5
cffi==1.9.1
click==6.6
coverage==4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-Migrate==2.0.2
flask-mongoengine==1.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
Flask-WTF==0.14.3
idna==3.1
itsdangerous==0.24
Jinja2==2.11.3
Mako==1.0.6
MarkupSafe==0.23
marshmallow==3.10.0
mongoengine==0.22.1
pycparser==2.20
PyJWT==1.4.2
pymongo==3.11.3
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==1.0.1
WTForms==2.3.3